added javadocs
