+++ src/main/java/com/hot2019pot/seattled41/xl/potluck/planner/PotluckUserController.java
@@ -15,25 +15,42 @@ import java.security.Principal;
 import java.sql.Date;
 import java.util.ArrayList;
 
-
+/**
+ * This is the Controller
+ * for a PotluckUser.
+ */
 @Controller
 public class PotluckUserController {
 
     @Autowired
-    PotluckUserRepository potLuckUserRepository;
+    private PotluckUserRepository potLuckUserRepository;
 
     @Autowired
-    PotluckRepository potLuckRepository;
+    private PotluckRepository potLuckRepository;
+
     @Autowired
-    PasswordEncoder passwordEncoder;
+    private PasswordEncoder passwordEncoder;
 
-
+    /**
+     * Get mapping to retrieve registration page
+     * for new user.
+     * @return String, html page
+     */
     @GetMapping("/signup")
     public String getsignup(){
         return "signup";
     }
 
 
+    /**
+     * Post mapping to save
+     * information of new user.
+     * @param firstname String, user first name
+     * @param lastname String, user last name
+     * @param username String, user username (must be unique)
+     * @param password String, user password
+     * @return View, html to access logged-in home
+     */
     @PostMapping("/signup")
     public RedirectView addNewUser( String firstname,String lastname,String username, String password){
         PotluckUser newUser=  new PotluckUser(firstname,lastname,username,passwordEncoder.encode(password));
@@ -43,38 +60,30 @@ public class PotluckUserController {
         return new RedirectView("/");
     }
 
-
+    /**
+     * Get mapping for user to route to
+     * page to add a new Potluck.
+     * @param m Model, hive/cache for front-facing
+     * @param p Principal, logged-in user object
+     * @return String, page to retrieve
+     */
     @GetMapping("/Potluck/add")
     public String createPotluck(Model m, Principal p){
         m.addAttribute("principal", p);
         return "createPotluck";
     }
 
-    @GetMapping("/home")
-    public String goHomePage(Model m, Principal p){
-        if (p ==null) {
-            m.addAttribute("principal", null);
-        }else {
-            m.addAttribute("principal", p);
-        }
-        return "home.html";
-    }
-
-    @GetMapping("/logout_complete")
-    public String getLogoutPage() {
-        return "logout_completed.html";
-    }
-
-    @GetMapping("/aboutus")
-    public String getAboutUs(Model m, Principal p) {
-        if (p ==null) {
-            m.addAttribute("principal", null);
-        }else {
-            m.addAttribute("principal", p);
-        }
-        return "aboutus.html";
-    }
-
+    /**
+     * Post mapping to create new Potluck object
+     * and save it to db.
+     * @param p Principal, logged-in user
+     * @param eventname String, Potluck name
+     * @param dateofPotluck Date, Potluck date
+     * @param location String, Potluck location
+     * @param details String, Potluck additional details
+     * @param m Model, hive/cache for front-facing
+     * @return String, html page to retrieve
+     */
     @PostMapping("/Potluck")
     public String newPotluck(Principal p, String eventname, Date dateofPotluck, String location, String details, Model m){
         PotluckUser creator = potLuckUserRepository.findByUsername(p.getName());
@@ -96,7 +105,6 @@ public class PotluckUserController {
         return  "redirect:/Potlucks/"+find.id;
     }
 
-
     private String generateCode(){
         // chose a Character random from this String
         String AlphaNumericString = "ABCDEFGHIJKLMNOPQRSTUVWXYZ"
@@ -107,7 +115,6 @@ public class PotluckUserController {
         StringBuilder sb = new StringBuilder(5);
 
         for (int i = 0; i < 5; i++) {
-
             // generate a random number between
             // 0 to AlphaNumericString variable length
             int index
@@ -118,9 +125,7 @@ public class PotluckUserController {
             sb.append(AlphaNumericString
                     .charAt(index));
         }
-
         return sb.toString();
-
     }
 
 }
